moved logics to service
added service saving
