+++ src/app/models/invoice.model.ts
@@ -1,0 +1,7 @@
+export interface Invoice {
+  id?: string;
+  clientName: string;
+  billingPeriod: string;
+  total: number;
+  setteled: boolean;
+}
