+++ src/assets/i18n/en.json
@@ -13,7 +13,7 @@
     "ACTIONS": "Actions",
     "DELETE": "Delete",
     "ADD SERVICE": "Add Service",
-    "DOWNLOAD": "Download",
+    "DOWNLOAD": "Save & Download",
     "TOTAL": "Total"
   }
 }
